requirements.txt
anytree==2.8.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
entrypoints==0.3
filelock==3.4.2
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
fonttools==4.28.5
greenlet==1.1.2
idna==3.3
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
jsonschema==4.3.2
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-http-over-ws==0.0.8
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.2
notebook==6.4.6
numexpr==2.7.3
numpy==1.20.3
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
regex==2021.4.4
requests==2.27.1
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.31
tables==3.6.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.61.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
virtualenv==20.13.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
zipp==3.6.0
gunicorn
